Update extractor usage for configurable configfile
